requirements.txt
aiogram>=3.4.1
yt-dlp>=2023.12.30
python-dotenv>=1.0.1
nest_asyncio
